Refactor (factory) code cleanup v2
