+++ src/Services/AdminService/AdminService.cs
@@ -1,6 +1,5 @@
 ﻿using Qubiz.QuizEngine.Database.Entities;
 using Qubiz.QuizEngine.Database.Repositories;
-using Qubiz.QuizEngine.Database;
 using Qubiz.QuizEngine.Infrastructure;
 using System;
 using System.Threading.Tasks;
